secure_notes_app/backend/requirements.txt
fastapi
uvicorn[standard]
boto3
PyJWT
requests
